requirements.txt
pyperclip
tqdm
python-docx
fpdf
